pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "fastapi-mail>=1.5.7",
    "psycopg-binary>=3.2.11",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
]
